pyproject.toml
[tool.poetry]
name = "mqtt-notifysend-bridge"
version = "0.1.0"
description = "Bridge between MQTT and Notify-Send"
authors = ["JamoDevNich"]
license = "LGPL-2.0"

[tool.poetry.dependencies]
python = "^3.6"
paho-mqtt = "^1.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry.scripts]
mqtt-notifysend-bridge = 'mqtt_notifysend_bridge:main'
